pyproject.toml
[project]
name = "fastai"
version = "0.1.0"
description = "FastAI Site Repository."
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "aiofiles>=24.1.0",
    "editorconfig-checker==3.2.1",
    "fastapi[standard]==0.115.12",
    "pydantic-settings>=2.10.1",
    "ruff==0.11.13",
]
